Created a class for the virtualized function
- Created a class for the virtualized function that
carries the argon module as a property (user doesn't
need to import modules used in the virtualized
function that they don't know about)

- Fixed a small error with deleting temporary
variables in the If node transformation

- Blocks now have the "inputs" field properly
implemented. This field captures all of the inputs
that are used in the statements of the block but
are not defined in the block itself.