Fixing srcctx issue with IfThenElse for ifs
Could not simply add on the lineno and col_offset info to the ast.Call
node for stage_if(). I tried doing so and propagating all those changes
down the tree for that node, but for some reason the call stack still
reflects a different line number. So instead, I am passing in the
file name as a parameter into the Transformer, and I am manually
creating the SrcCtx without looking at the stack frames anymore.